✅ test: relocate ParametricQuantity-specific tests to unxts.parametric
Move the tests that exercise ParametricQuantity out of the core unxt
suite and into the parametric package, so parametric behavior is tested
where it lives:

- test_interchangeability.py (was tests/unit/test_quantity_interchangeability)
- test_plum_faithful.py (was tests/integration/test_plum)
- test_printing.py, test_parametric_quantity.py, test_static_value.py
  (extracted from tests/unit/test_quantity_printing, test_quantity,
  test_static_quantity)

Core tests keep only what concerns the default Quantity / StaticQuantity;
where they reference PQ they now import it from unxts.parametric. The
stack-with-dimensionless test now expects a plain Quantity (matching the
decoupled Angle/stack rules). Update the moved-test per-file ruff ignore
path in pyproject.toml.
