+++ tests/integration/quaxed/test_numpy.py
@@ -13,6 +13,7 @@ from jax._src.numpy.setops import (
 )
 from jax.numpy import iinfo as IInfo  # noqa: N812
 from numpy import finfo as FInfo  # noqa: N812
+from unxts.parametric import PQ
 
 import quaxed.lax as qlax
 import quaxed.numpy as jnp
@@ -978,11 +979,11 @@ def test_positive():
 def test_pow_quantity_power():
     """Test `pow` with a Quantity exponent.
 
-    A Quantity exponent must be a *parametric* dimensionless quantity
-    (``u.PQ``): that is the signature the ``pow`` primitive registers.
+    A dimensionless parametric quantity (``PQ``) is a valid exponent; the
+    parametric-exponent ``pow`` rule lives in ``unxts.parametric``.
     """
     x = u.Q(jnp.asarray([1, 2, 3], dtype=float), "m")
-    y = u.PQ(jnp.asarray(4, dtype=float), "")
+    y = PQ(jnp.asarray(4, dtype=float), "")
     got = jnp.pow(x, y)
     exp = u.Q(jnp.pow(x.value, y.value), "m4")
 
@@ -1394,14 +1395,14 @@ def test_stack_array_with_dimensionless():
     """Test `stack` with arrays and dimensionless quantities.
 
     Stacking a leading raw array with a dimensionless quantity promotes the
-    result to a ``ParametricQuantity`` (``u.PQ``); see ``stack_p_v``.
+    result to a plain ``Quantity``; see ``stack_p_v``.
     """
     x = jnp.asarray([1, 2, 3], dtype=float)
     y = u.Q(jnp.asarray([4, 5, 6], dtype=float), "")
     got = jnp.stack((x, y))
     exp = u.Q(jnp.stack((x, y.value)), "")
 
-    assert isinstance(got, u.PQ)
+    assert isinstance(got, u.Q)
     assert str(got.unit) == ""
     assert jnp.array_equal(got.value, exp.value)
